# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rewaa Quran Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  level?: string | null;
  /** @nullable */
  teacherId?: number | null;
  /** @nullable */
  teacherName?: string | null;
  role: string;
  createdAt?: Date;
}