# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rewaa Quran Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  name: string;
  email?: string;
  phone?: string;
  password: string;
  age?: number;
  level?: string;
  /** @nullable */
  teacherId?: number | null;
}